Refactoring frontend
