--- frontend/src/Login.css
@@ -1,55 +1,0 @@
-/* Remove default margins and paddings from body */
-html, body {
-    margin: 0;
-    padding: 0;
-    flex-direction: column;
-    align-items: center;
-}
-
-label {
-    font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
-    display: block;
-    font-size: 1.17em;
-    margin-top: 1em;
-    margin-bottom: 1em;
-    margin-left: 0;
-    margin-right: 0;
-    font-weight: bold;
-}
-
-#root {
-    margin: 0;
-    width: auto;
-    min-width: 1100px;
-    min-height: 800px;    
-    flex-direction: column;
-    align-items: center;
-}
-
-.login-container {
-	margin: auto;
-    max-width: fit-content;
-    max-height: fit-content;
-    text-align: left;
-    margin-top: 15%;
-    padding: 30px 50px;
-    background-color: #CA3434;
-    border-radius: 8px;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
-
-.button-container {
-    margin-top: 30px;
-    text-align: center;
-}
-
-.login-button {
-    border-style: double;
-    border-color: white;
-    background-color: #CA3434;
-    border-radius: 8px;
-    width: auto;
-}
