+++ frontend/src/index.css
@@ -14,6 +14,10 @@
   max-width: fit-content;
   margin-left: auto;
   margin-right: auto;
+
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
 }
 
 a {
@@ -29,8 +33,7 @@ body {
   margin: 0;
   display: flex;
   place-items: center;
-  min-width: 320px;
-  min-height: 100vh;
+
 }
 
 h1 {
