+++ frontend/src/main.jsx
@@ -2,20 +2,22 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
 import './index.css'
-import Profile from './Profile.jsx'
-import Signup from './Signup.jsx'
+import './styles/global.css'
+import Profile from './pages/Profile.jsx'
+import Signup from './pages/Signup.jsx'
 //import App from './App.jsx'
-import Recomendations from './Recomendations'
-import Login from "./Login.jsx"
+import Recommendations from './components/Recommendations.jsx'
+import Home from './pages/Home.jsx'
+import Login from "./pages/Login.jsx"
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <BrowserRouter>
       <Routes>
-        <Route path="/" element={< Recomendations />} />
-        <Route path="/login" element={<Login/>} />
+        <Route path="/" element={< Home />} />
+        {/* <Route path="/login" element={<Login/>} />
         <Route path="/signup" element={<Signup/>} />
-        <Route path="/profile" element={<Profile/>} />
+        <Route path="/profile" element={<Profile/>} /> */}
       </Routes>
     </BrowserRouter>
   </StrictMode>
