+++ frontend/src/styles/Profile.module.css
@@ -1,0 +1,66 @@
+/* Remove default margins and paddings from body */
+html, body {
+    margin: 0;
+    padding: 0;
+    flex-direction: column;
+    align-items: center;
+}
+
+#root {
+    margin: 0;
+    text-align: left;
+    width: auto;
+    min-width: 1100px;
+    min-height: 800px;
+}
+
+.profile-container {
+    width: auto;
+}
+
+.horizontal-layout {
+    width: auto; 
+    overflow: hidden;
+}
+
+.pic-and-linked-accounts {
+    min-height: 800px;
+    width: 30%;
+    float: left;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+    background-color: #CA3434;
+    margin: 3px;
+    border-radius: 8px;
+}
+
+.vertical-layout {
+    width: 70%;
+    margin-top: 0;
+    margin-left: 30%;
+    flex-direction: column;
+    align-items: center;
+}
+
+.played-games {
+    min-height: 150px;
+    background-color: #CA3434;
+    margin: 0px 6px 0px 10px;
+    border-radius: 8px;
+    padding: 0px 10px 10px 10px;
+}
+
+.liked-games {
+    min-height: 150px;
+    background-color: #CA3434;
+    margin: 0px 6px 0px 10px;
+    border-radius: 8px;
+    padding: 0px 10px 10px 10px;
+}
+
+.profile-pic {
+    border-radius: 50%;
+    width: 75%;
+    height: auto;
+}
